requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pillow==10.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.2.1
pydantic_core==2.14.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.0.134
urllib3==2.2.1
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
